+++ src/actions/index.js
@@ -1,2 +1,6 @@
-export * from "./action_post";
-export * from "./action_get";
+export * from "./action_hospitals";
+export * from "./action_floors";
+export * from "./action_rooms";
+export * from "./action_beds";
+export * from "./action_patients";
+export * from "./action_data";
