+++ src/constants/ActionTypes.js
@@ -7,21 +7,32 @@ export const URL = "10.94.71.150";
 export const PORT = 3000;
 export const ROOT_URL = `http://${URL}:${PORT}`;
 
-// get request
+// Hospital get request
 export const FETCH_HOSPITALS = "fetch_hospitals";
 export const FETCH_HOSPITAL = "fetch_hospital";
+
+// Hospital post request
+export const ADD_HOSPITAL_INFO = "add_hospital_info";
+export const EIDT_HOSPITAL_INFO = "edit_hospital_info";
+
+// Floor get request
 export const FETCH_FLOORS_AT = "fetch_floors_at";
 export const FETCH_FLOOR = "fetch_floor";
+
+// Room get request
 export const FETCH_ROOMS_AT = "fetch_rooms_at";
 export const FETCH_ROOM = "fetch_room";
+
+// Bed get request
 export const FETCH_BEDS_AT = "fetch_beds_at";
 export const FETCH_BED = "fetch_bed";
+
+// Patient get request
 export const FETCH_PATIENTS = "fetch_patients";
 export const FETCH_PATIENT = "fetch_patient";
 export const FETCH_PATIENTS_SEACHED = "fetch_patients_searched";
+
+// Data get request
 export const FETCH_SENSOR_DATA = "fetch_sensor_data";
 
-// post request
-export const ADD_HOSPITAL_INFO = "ADD_hospital_info";
-export const EIDT_HOSPITAL_INFO = "edit_hospital_info";
